Add basic routing
This adds the core-life module which wraps the home, work, restaurant, and gym components.

The app.routes.ts file configures the Router which is imported in the app module. The <router-outlet> is in the app.component.html file which displays these routes.

Bootstrap has been added to the project. The css file is imported in the angular-cli.json file (in the styles array).